+++ src/api2/api-caller.js
@@ -180,10 +180,16 @@ export class ApiCaller {
    */
   async _call (opts) {
     let config = {
+      baseURL: this._baseURL,
       params: opts.query || {},
+      paramsSerializer: function (params) {
+        return Object.entries(params)
+          .map(([key, value]) => `${key}=${encodeURIComponent(value)}`)
+          .join('&')
+      },
       data: opts.data || {},
       method: opts.method,
-      url: this._baseURL + opts.endpoint // TODO: smartly build url
+      url: opts.endpoint // TODO: smartly build url
     }
 
     config = middlewares.flattenToAxiosJsonApiQuery(config)