+++ CHANGELOG.md
@@ -8,6 +8,11 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
 ## [Unreleased]
+### Added
+- `ReviewRequestOpExt` attribute to the `ReviewRequestBuilder._prepareArgs` method
+
+### Fixed
+- Using account role instead of account type in `ReviewRequestBuilder.reviewLimitsUpdateRequest` method
 
 ### Removed
 - All tasks from createSaleCreationRequest
@@ -19,6 +24,20 @@ for further information about branching and tagging conventions.
 - accountRole field enveloping in setFees operation
 - do not throw error if cannot parse errors in BadRequestError
 
+## [1.0.0-x.17] - 2019-02-23
+### Fixed
+- Setting 0 create_asset_tasks when no value provided
+
+## [1.0.0-x.16] - 2019-02-23
+### Fixed
+- `blobs` resource now work with `/account/:id/blobs` endpoint instead of `/users/:id/blobs`
+- `documents` resource now work with `/account/:id/documents` endpoint instead of `/users/:id/documents`
+- `api.wallets.create` and `api.wallets.recovery` now work with `/v3` horizon endpoints
+- `wallets.create` now building the parameters in valid JSON API format
+- AccountRole field enveloping in setFees operation
+- Do not throw error if cannot parse errors in BadRequestError
+- Fixed invalid signature on fetching response links
+
 ## [1.0.0-x.15] - 2019-02-21
 ### Added
 - `JsonapiResponse` exporting
@@ -29,6 +48,7 @@ for further information about branching and tagging conventions.
 
 ### Fixed
 - `makeChangeSignerTransaction` function according to the signer roles
+- Serializing of the axios params before submitting
 
 ## [1.0.0-x.14] - 2019-02-21
 ### Added
@@ -172,7 +192,9 @@ for further information about branching and tagging conventions.
 ### Fixed
 - Change password method
 
-[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.0.0-x.15...HEAD
+[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.0.0-x.17...HEAD
+[1.0.0-x.17]: https://github.com/tokend/new-js-sdk/compare/1.0.0-x.16...1.0.0-x.17
+[1.0.0-x.16]: https://github.com/tokend/new-js-sdk/compare/1.0.0-x.15...1.0.0-x.16
 [1.0.0-x.15]: https://github.com/tokend/new-js-sdk/compare/1.0.0-x.14...1.0.0-x.15
 [1.0.0-x.14]: https://github.com/tokend/new-js-sdk/compare/1.0.0-x.13...1.0.0-x.14
 [1.0.0-x.13]: https://github.com/tokend/new-js-sdk/compare/1.0.0-x.12...1.0.0-x.13